Move menu.py and scoping.py into a new commands/helpers/ package
Neither one registers a CommandHandler — they're shared Telegram-aware
plumbing every command handler pulls from, not commands themselves.
Splitting them out keeps the flat commands/ listing to actual handlers.
Pure move, no behavior change.

Part of #58
