+++ src/ley_shards_bot/commands/helpers/scoping.py
